server/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.5
langchain-core==0.3.13
langchain-openai==0.2.4
langchain-text-splitters==0.3.1
langsmith==0.1.137
markdown-it-py==3.0.0
markdown_pdf==1.3
mdurl==0.1.2
multidict==6.1.0
numpy==1.26.4
openai==1.52.2
orjson==3.10.10
packaging==24.1
pocketbase==0.12.3
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
python-dotenv==1.0.1
python-multipart==0.0.16
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
uuid==1.30
uvicorn==0.32.0
validators==0.34.0
yarl==1.17.0
